requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
multidict==4.7.6
netifaces==0.10.6
Pillow==7.1.2
python-engineio==3.12.1
python-socketio==4.5.1
requests==2.23.0
twitchio==1.1.0
urllib3==1.25.9
websocket==0.2.1
websocket-client==0.57.0
websockets==8.1
yarl==1.4.2
